+++ platform/darwin/src/MGLVectorTileSource_Private.h
@@ -8,7 +8,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NS_SET_OF(NSString *) *)mapboxStreetsLanguages;
 
-+ (NSString *)preferredMapboxStreetsLanguage;
++ (nullable NSString *)preferredMapboxStreetsLanguage;
 
 - (NS_DICTIONARY_OF(NSString *, NSString *) *)localizedKeysByKeyForPreferredLanguage:(nullable NSString *)preferredLanguage;